readify_agi/requirements.txt
# Web???API
fastapi>=0.100.0,<1.0.0
uvicorn>=0.35.0,<1.0.0

# ???
sqlalchemy>=1.4.0,<2.0.0
aiomysql>=0.2.0,<1.0.0

# ????
python-dotenv>=1.0.0,<2.0.0

# ???????
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# HTTP???
starlette>=0.27.0,<1.0.0
typing-extensions>=4.5.0,<5.0.0
email-validator>=2.0.0,<3.0.0
httpx>=0.24.0,<1.0.0

# ??
cryptography>=40.0.0,<50.0.0

# AI?????
numpy>=1.24.0,<2.0.0
llama-parse>=0.6.0,<1.0.0
langchain>=0.1.5,<1.0.0
langchain-core>=0.1.14,<1.0.0
langchain-community>=0.1.5,<1.0.0
langchain-openai>=0.1.7,<1.0.0
pymilvus>=2.4.0,<2.6.0
openai>=1.3.0,<2.0.0
tiktoken>=0.9.0,<1.0.0
agently>=3.5.0,<4.0.0

# ??
pytest>=7.0.0,<9.0.0
pytest-asyncio>=0.21.0,<2.0.0

# ????
python-multipart>=0.0.20,<1.0.0  # ??????
pypdf>=5.0.0,<6.0.0  # ??PDF??
minio>=7.2.0,<8.0.0  # ????

# ??API
google-search-results>=2.4.0,<3.0.0  # SerpAPI ?
nacos-sdk-python>=3.0.2  # Nacos ???
PyYAML>=6.0,<7.0
